Move isRecordableOffence function into root lib folder
