+++ packages/core/lib/isRecordableOffence.test.ts
@@ -1,4 +1,4 @@
-import type { Offence } from "../../types/AnnotatedHearingOutcome"
+import type { Offence } from "../types/AnnotatedHearingOutcome"
 
 import isRecordableOffence from "./isRecordableOffence"
