+++ packages/core/lib/triggers/TRPS0011.test.ts
@@ -1,14 +1,14 @@
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
-import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
 import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
 import generatePncUpdateDatasetFromOffenceList from "../../phase2/tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
 import Phase from "../../types/Phase"
 import { PncOperation } from "../../types/PncOperation"
+import isRecordableOffence from "../isRecordableOffence"
 import isResultCompatibleWithDisposal from "../isResultCompatibleWithDisposal"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 import TRPS0011 from "./TRPS0011"
-jest.mock("../../phase2/lib/isRecordableOffence")
+jest.mock("../isRecordableOffence")
 jest.mock("./hasCompletedDisposal")
 jest.mock("../isResultCompatibleWithDisposal")
