+++ packages/core/lib/triggers/TRPS0013.test.ts
@@ -1,13 +1,13 @@
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
-import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
 import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
 import generatePncUpdateDatasetFromOffenceList from "../../phase2/tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
 import Phase from "../../types/Phase"
 import { PncOperation } from "../../types/PncOperation"
+import isRecordableOffence from "../isRecordableOffence"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 import TRPS0013 from "./TRPS0013"
-jest.mock("../../phase2/lib/isRecordableOffence")
+jest.mock("../isRecordableOffence")
 jest.mock("./hasCompletedDisposal")
 
 const mockedIsRecordableOffence = isRecordableOffence as jest.Mock
