+++ packages/core/lib/triggers/TRPS0013.ts
@@ -2,9 +2,9 @@ import TriggerCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/
 
 import type { TriggerGenerator } from "../../types/TriggerGenerator"
 
-import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
 import Phase from "../../types/Phase"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
+import isRecordableOffence from "../isRecordableOffence"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 
 const triggerCode = TriggerCode.TRPS0013
