+++ packages/core/phase2/exceptions/checkResultClassExceptions.ts
@@ -1,7 +1,7 @@
 import type { AnnotatedHearingOutcome, Offence, Result } from "../../types/AnnotatedHearingOutcome"
 
+import isRecordableOffence from "../../lib/isRecordableOffence"
 import isRecordableResult from "../../lib/isRecordableResult"
-import isRecordableOffence from "../lib/isRecordableOffence"
 
 type CheckExceptionFn = (offence: Offence, result: Result, offenceIndex: number, resultIndex: number) => void
