+++ packages/core/phase2/lib/generateOperations/generateOperations.ts
@@ -1,11 +1,11 @@
 import type { AnnotatedHearingOutcome } from "../../../types/AnnotatedHearingOutcome"
 import type { ResultClassHandler } from "./resultClassHandlers/ResultClassHandler"
 
+import isRecordableOffence from "../../../lib/isRecordableOffence"
 import isRecordableResult from "../../../lib/isRecordableResult"
 import { PncOperation } from "../../../types/PncOperation"
 import { type Operation } from "../../../types/PncUpdateDataset"
 import ResultClass from "../../../types/ResultClass"
-import isRecordableOffence from "../isRecordableOffence"
 import deduplicateOperations from "./deduplicateOperations"
 import filterDisposalsAddedInCourt from "./filterDisposalsAddedInCourt"
 import { handleAdjournment } from "./resultClassHandlers/handleAdjournment"
